fix read stuck on SIM800L
The +CIPRXGET response indicates the amount of bytes that are ready to be read (the 3rd parameter). modemreadassumes that indeed these amout of characters CAN be read. This turns out not to be case. Sometimes stream.readreturns -1before the expected number of characters are read!

remove the wait when no data is available and
when character read return -1 (no more chars to read), quit the routine.