MKR GSM 1400 optimal speed (with Events)
Added an Events (Callback) class to allow the sketch to perform specific actions, such as PowerOn and ModemReady. Both callbacks perform specific MKR GSM 1400 actions, that are different per Modem.

The events code (*16 bytes* of additional code) is an elegant way to avoid to split methods and perform optional code (not all Modem will be put to a higher speed). Most modems do need to be turned on (PowerOn event/callback)