[WIP] Use relative includes
So that this library can reside in the sketch directory under src/ rather than in the global libraries folder. This allows one to have a dedicated version of this library in use by the sketch, especially in cases where multiple versions of the library is available. This also allows one to have a specific version committed to git along with the sketch.