TinyGsmClientSequansMonarch.h: use data mode
Use data mode instead of text mode to be able to send any content such
as raw binary. This will allow us to use TinyGSM with ArduinoBearSSL and
ArduinoMqtt.

It should be noted that for an unclear reason GSM_NL must be set to '\n'
instead of '\r\n' to avoid "CME ERROR: operation not supported" in this
mode with the Monarch GMS01Q.

Signed-off-by: Fabrice Fontaine <fabrice.fontaine@orange.com>